Implement q_insert_head
First check for null queue. Then allocate memory for new element with
null pointer check. Finally use test_strdup function to allocate and
copy the new string into the element safely.

The original version is implemented with strcpy, which is not safe
because it does not check buffer lengths and may very well overwrite
memory zone contiguous to the intended destination.

Change-Id: I8b7de46f4c2710dc58d0c3e04f5ac8fc01f3ea17